feat(payment): PAYPAL-3602 renamed PayPalCommerceConnectTracker to PayPalCommerceAnalyticTracker + added Fastlane support
