+++ packages/core/src/analytics/paypal-commerce-connect-tracker/paypal-commerce-analytic-tracker-service.ts
@@ -1,4 +1,4 @@
-export default interface PayPalCommerceConnectTrackerService {
+export default interface PayPalCommerceAnalyticTrackerService {
     customerPaymentMethodExecuted(): void;
     paymentComplete(): void;
     selectedPaymentMethod(methodId: string): void;
